rename: edit acoount app to user app
- Renamed account app name to user app due to duplication error with all auth settings
